fix(davinci): lower JSX v-show through S2
